% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_analysis_fxns.R
\name{all_max_cum_infect}
\alias{all_max_cum_infect}
\title{Get all final cumulative infected numbers from sims}
\usage{
all_max_cum_infect(trials)
}
\arguments{
\item{trials}{A list with simulated zika outbreaks. Usually utput of run_n_zika_sims()}
}
\value{
A vector of all the max cumulatively infected individuals for each simulation.
Calls \code{\link{last_cuminfect_local}}
}
\description{
Get all final cumulative infected numbers from sims
}
\examples{

}
