% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_epi_prob_by_d}
\alias{county_epi_prob_by_d}
\title{Epidemic probability for counties}
\format{A dataframe with 1332 rows and 5 variables:
\describe{
  \item{r_not}{Texas county r0}
  \item{disc_prob}{daily probability of being reported in simulation}
  \item{intro_rate}{daily rate of introductions for simulation}
  \item{detected}{Number of autochthonous reported cases}
  \item{prob_epidemic}{probability of an epidemic give the number of reported cases}
  ...
}}
\source{
county simulation data
}
\usage{
county_epi_prob_by_d
}
\description{
Figure 4b data
}
\keyword{datasets}

