% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_prob_detect_x}
\alias{county_prob_detect_x}
\title{County Detection probability}
\format{A dataframe with 1224 rows and 5 variables:
\describe{
  \item{r_not}{Texas county r0}
  \item{disc_prob}{daily probability of being reported in simulation}
  \item{intro_rate}{daily rate of introductions for simulation}
  \item{detected}{Number of autochthonous reported cases}
  \item{prob_detect}{probability detecting the number of detected cases}
  ...
}}
\source{
county simulation data
}
\usage{
county_prob_detect_x
}
\description{
Figure 4a data - probability of observing x detected cases in county
}
\keyword{datasets}
