% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_risk_data}
\alias{county_risk_data}
\title{County Risk spreadsheet}
\format{A dataframe with 254 rows and 24 variables:
\describe{
  \item{Metro}{Metro area that county is associated with (integer)-not used}
  \item{id}{Texas County ID}
  \item{Geography}{County name format 'county_name "County, Texas"'}
  \item{GDP}{County GDP}
  \item{importation_probability}{Probability that next imported case occurs in that county}
  \item{importation.current}{Importation rate Quarter 1 for county}
  \item{importation.projected}{Projected county importation rate for Quarter 3}
  \item{importation.worse.projected}{Worst case county importation projection for Q3}
  \item{mosquito.abundance}{Mosquito abundance for county from Kraemer 2015 global distribution data}
  \item{rnott.expected.round}{Expected august r0 for county}
  \item{low.round}{Low GDP effect r0 for august}
  \item{high.round}{High GDP effect r0 for august}
  \item{hetero.round}{Heterogeneous GDP effect for r0 for august}
  \item{temperature}{Historic average county temperature for August}
  \item{eip}{Extrinsic incubation period for mosquitoes at average temperature}
  ...
}}
\source{
Zika epidemic simulations with R0=1.1 or high risk group, intro rate = 0.01 or high risk group
}
\usage{
county_risk_data
}
\description{
Data for importation/transmission risk for all counties
}
\keyword{datasets}
