% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_analysis_fxns.R
\name{cum_detect_local}
\alias{cum_detect_local}
\title{Returns cumulative vector of locally detected cases}
\usage{
cum_detect_local(x)
}
\arguments{
\item{x}{a single trial run}
}
\value{
Cumulative detectections in vector that are local
}
\description{
Returns cumulative vector of locally detected cases
}
\examples{

}

