% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_analysis_fxns.R
\name{cumcases_by_detects}
\alias{cumcases_by_detects}
\title{Cumulative cases for detects for single simulation run}
\usage{
cumcases_by_detects(df, max_detect)
}
\arguments{
\item{df}{A dataframe that is the result of a single zika simulation run}

\item{max_detect}{the maximum number of detections desired to be analyzed}
}
\value{
the maximum cumulative cases and max prevalence for each unique detection
}
\description{
Cumulative cases for detects for single simulation run
}
