% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epi_prob_by_reported_data}
\alias{epi_prob_by_reported_data}
\title{Prevalence by reported cases data}
\format{A dataframe with 404 rows and 6 variables:
\describe{
  \item{risk_level}{risk level for simulation "High Risk" or "1.1"}
  \item{disc_prob}{reporting rate of simulation}
  \item{detected}{Number of reported autochthonous cases}
  \item{value}{probability of an epidemic given you've detected x cases}
  ...
}}
\source{
Zika epidemic simulations with R0=1.1 or high risk group, intro rate = 0.01 or high risk group
}
\usage{
epi_prob_by_reported_data
}
\description{
Figure 2b data
}
\keyword{datasets}

