% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_dying_epidemic_data}
\alias{ex_dying_epidemic_data}
\title{Dying outbreak example}
\format{A long format data frame with 172 rows and 3 variables:
\describe{
  \item{time}{day of outbreak simulation}
  \item{variable}{specifies asymptomatic/symptomatic}
  \item{value}{number of symptomatic versus asymptomatic individuals}
  ...
}}
\source{
Zika epidemic simulations with R0=1, reporting rate=0.011, intro rate = 0.1, 15% symptomatic
}
\usage{
ex_dying_epidemic_data
}
\description{
Figure 1 data
}
\keyword{datasets}

