% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_outbreak_sims}
\alias{ex_outbreak_sims}
\title{Simulated stochastic outbreak data}
\format{A data frame with 1108351 rows and 15 variables:
\describe{
  \item{risk_level}{risk level for simulation "High Risk" or "1.1"}
  \item{disc_prob}{reporting rate of simulation}
  \item{.id}{Simulation number used to separate individual outbreaks}
  \item{time}{.id specific time of outbreak simulation}
  \item{New_Exposed}{Newly exposed individuals}
  \item{New_Infection}{Newly infectious individuals}
  \item{Total_Infections}{Total number of infectious individuals}
  \item{Cumulative_Infections}{Cumulative number of infectious individuals}
  \item{New_Intro}{Newly imported (introduced) infectious cases}
  \item{Total_Intro_Infections}{Total number of imported (introduced) infectious cases}
  \item{Cumulative_Intro_Infections}{Cumulative number of imported (introduced) infectious cases}
  \item{New_Detections}{Newly reported (detected) cases}
  \item{Cum_Detections}{Cumulative number of detected cases for simulation}
  \item{New_Intro_Detections}{Newly reported (detected) cases that were initially imported}
  \item{Cum_Intro_Detections}{Cumulative reported (detected) cases that were initially imported}
  ...
}}
\source{
Zika epidemic simulations with R0=1.1 or high risk group, intro rate = 0.01 or high risk group
}
\usage{
ex_outbreak_sims
}
\description{
Figure 2a data
}
\keyword{datasets}
