% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_analysis_fxns.R
\name{freq_above_thresh}
\alias{freq_above_thresh}
\title{Get frequency cases above cum threshold and prev threshold for unique detected}
\usage{
freq_above_thresh(df, detected, cum_threshold, prev_threshold, num_necessary)
}
\arguments{
\item{df}{dataframe formatted output from \code{\link{get_cumcases_by_detects_all}}. Has one columns for detected, cum_infections, and max_prevalence.}

\item{detected}{single integer number of interest}

\item{cum_threshold}{cumulative threhold for epidemic classification}

\item{prev_threshold}{prevalence threshold for epidemic classification}

\item{num_necessary}{number of trials necessary for epi classification usually 1\% of trials.}
}
\value{
a single probability of epidemic for single detected value.
}
\description{
Returns the frequency that cases were found to be above a threshold.
}
