% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_analysis_fxns.R
\name{freq_above_thresh_vec}
\alias{freq_above_thresh_vec}
\title{Vectorizes the freq_above_thresh for the number of detected}
\usage{
freq_above_thresh_vec(df, detected, cum_threshold, prev_threshold,
  num_necessary)
}
\arguments{
\item{df}{dataframe formatted output from \code{\link{get_cumcases_by_detects_all}}. Has one columns for detected, cum_infections, and max_prevalence.}

\item{detected}{vector of detected of interest}

\item{cum_threshold}{cumulative threhold for epidemic classification}

\item{prev_threshold}{prevalence threshold for epidemic classification}

\item{num_necessary}{number of trials necessary for epi classification usually 1\% of trials.}
}
\value{
A vector of epidemic probs for the given detected values.
}
\description{
Called from the get_epidemic_prob_by_d
}
