% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_analysis_fxns.R
\name{get_cumcases_by_detects_all}
\alias{get_cumcases_by_detects_all}
\title{Get all cumcases and prevalence by detects}
\usage{
get_cumcases_by_detects_all(trials, max_detect)
}
\arguments{
\item{trials}{a list of zika trial simulations}

\item{max_detect}{the maximum amount of detections interested in for analysis}
}
\value{
a single probability of epidemic for single detected value
}
\description{
Vectorized version of \code{\link{cumcases_by_detects}}, to get all trial information.
}
