% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_analysis_fxns.R
\name{get_epidemic_prob_by_d}
\alias{get_epidemic_prob_by_d}
\title{Get the epidemic probability as a function of number of reported cases}
\usage{
get_epidemic_prob_by_d(trials, prev_threshold, cum_threshold, max_detect = 50,
  num_necessary = 1)
}
\arguments{
\item{trials}{A list with simulated zika outbreaks. Usually utput of \code{\link{run_n_zika_sims}}}

\item{prev_threshold}{The maximum autochthonous prevalence necessary to be classifed as an epidemic - depends on individual scenario run values}

\item{cum_threshold}{The cumulative autochthonous infections necessary to be classified as an epidemic - usually the e_thresh value of the runs}

\item{max_detect}{The maximum number of detections to go to.}

\item{num_necessary}{Number of instances of trials to be necessary before it gets a probability. Usually want to be ~1% of total runs}
}
\value{
A dataframe of rows max_detect+1 that has column for detected, and column for prob_epidemic given that detected number
}
\description{
Get the epidemic probability as a function of number of reported cases
}
\examples{
get_epidemic_prob_by_d(trials, 5, 100, 15, 1)

}

