% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growing_arrows}
\alias{growing_arrows}
\title{Growing outbreak arrow data}
\format{A long format data frame with 18 rows and 3 variables:
\describe{
  \item{time}{day of outbreak simulation}
  \item{yval}{y value for number total infectious at the given time}
  \item{type}{either import or report to determine arrows and lines}
  ...
}}
\source{
Zika epidemic simulations with R0=1, reporting rate=0.011, intro rate = 0.1, 15% symptomatic
}
\usage{
growing_arrows
}
\description{
Figure 1 data
}
\keyword{datasets}
