% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_analysis_fxns.R
\name{last_cuminfect_local}
\alias{last_cuminfect_local}
\title{Get single final cumulative infected numbers from sims}
\usage{
last_cuminfect_local(x)
}
\arguments{
\item{x}{a single trial run}
}
\value{
the max cumulative infected individuals for each simulation
}
\description{
Get single final cumulative infected numbers from sims
}
\examples{

}
