% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{major_city_loc}
\alias{major_city_loc}
\title{Texas Major city location}
\format{A dataframe with 15 rows and 4 variables:
\describe{
  \item{ID}{Texas city ID number}
  \item{lat}{latitude of city}
  \item{lon}{longitude of city}
  \item{Name}{Name of city}
  ...
}}
\source{
google maps
}
\usage{
major_city_loc
}
\description{
Data for location of major texas cities
}
\keyword{datasets}

