% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_analysis_fxns.R
\name{max_local_prev}
\alias{max_local_prev}
\title{Get single maximum of local prevalence}
\usage{
max_local_prev(x)
}
\arguments{
\item{x}{a single trial run}
}
\value{
the maximum local prevalence from run
}
\description{
Get single maximum of local prevalence
}
\examples{

}

