% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fxns.R
\name{plot_epi_prob}
\alias{plot_epi_prob}
\title{Plot epidemic risk}
\usage{
plot_epi_prob(df, max_detect = 10)
}
\arguments{
\item{df}{A dataframe that is the output of \code{\link{get_epidemic_prob_by_d}}}

\item{max_detect}{The maximum number of cases to plot on the x-axis}
}
\value{
A ggplot object with cowplot theme final size histogram
}
\description{
Plots the probability of an epidemic as a function of the number of reported local cases
}
\examples{
travis_parms <- get_county_parms("travis")
travis_sims <- run_n_zika_sims(num_reps = 1000, sim_parms)
travis_epi_prob <- get_epidemic_prob_by_d(travis_sims, prev_threshold = 10, cum_threshold = 100)
plot_epi_prob(travis_epi_prob)
}
