% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_fxns.R
\name{plot_fig1}
\alias{plot_fig1}
\title{Plot Fig 1 from Zika risk assessment}
\usage{
plot_fig1(panels = NULL)
}
\arguments{
\item{panels}{Optional character argument ("a", "b") specifying individual desired panels to return. Accepts uppercase letters as well.}
}
\value{
A ggplot object with cowplot theme that contains both figure panels
}
\description{
Plot Fig 1 from Zika risk assessment
}
\examples{
plot_fig1()
}

