% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_fxns.R
\name{plot_fig2}
\alias{plot_fig2}
\title{Plot Fig 2 from Zika risk assessment}
\usage{
plot_fig2(panels = NULL)
}
\arguments{
\item{panels}{Optional character argument ("a", "b", "c") specifying individual desired panels to return. Accepts uppercase letters as well.}
}
\value{
A ggplot object for specified plot
}
\description{
Plot Fig 2 from Zika risk assessment
}
\examples{
plot_fig2()
plot_fig2(panels="b")
}
