% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_fxns.R
\name{plot_fig3}
\alias{plot_fig3}
\title{Plot Fig 3 from Zika risk assessment}
\usage{
plot_fig3(panels = NULL)
}
\arguments{
\item{panels}{Optional character argument ("a", "b") specifying individual desired panels to return. Accepts uppercase letters as well.}
}
\value{
A ggplot object for specified plot
}
\description{
Plot Fig 3 from Zika risk assessment
}
\examples{
plot_fig3()
plot_fig3(panels="a")
}
