% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_fxns.R
\name{plot_fig4}
\alias{plot_fig4}
\title{Plot Fig 4 from Zika risk assessment}
\usage{
plot_fig4(panels = NULL)
}
\arguments{
\item{panels}{Optional character argument ("a", "b", "c") specifying individual desired panels to return. Accepts uppercase letters as well.}
}
\value{
A ggplot object for specified plot
}
\description{
This can take a long time to run, because the spatial plotting in ggplot can be slow.
}
\examples{
plot_fig4()
plot_fig4(panels="a")

# It's suggested that you save plot:
fig4 <- plot_fig4()
save_plot("fig4.pdf", fig4, base_aspect_ratio = 3, base_height=5)
}

