% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fxns.R
\name{plot_final_sizes}
\alias{plot_final_sizes}
\title{Plot histogram of the final sizes of zika outbreaks}
\usage{
plot_final_sizes(trials)
}
\arguments{
\item{trials}{A list with simulated zika outbreaks. Usually utput of run_n_zika_sims()}
}
\value{
A ggplot object with cowplot theme final size histogram
}
\description{
Plot histogram of the final sizes of zika outbreaks
}
\examples{
sims <- run_n_zika_sims(100, zika_def_parms())
plot_final_sizes(sims)
}

