% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fxns.R
\name{plot_zika_outbreaks}
\alias{plot_zika_outbreaks}
\title{Plot outbreaks}
\usage{
plot_zika_outbreaks(trials, cases = "local")
}
\arguments{
\item{trials}{A list with simulated zika outbreaks. Usually utput of run_n_zika_sims()}

\item{cases}{A string dictating what to plot c("local", "total", "all")}
}
\value{
A ggplot object with cowplot theme final size histogram
}
\description{
Plots the cumulative autochthonously infected individuals for all outbreaks simulated
}
\examples{
sims <- run_n_zika_sims(100, zika_def_parms())
plot_zika_outbreaks(sims)
}

