% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prev_by_reported_data}
\alias{prev_by_reported_data}
\title{Prevalence by reported cases data}
\format{An object of class \code{data.frame} with 404 rows and 6 columns.}
\source{
Zika epidemic simulations with R0=1.1 or high risk group, intro rate = 0.01 or high risk group
}
\usage{
prev_by_reported_data
}
\description{
Figure 2b data - Gives the median and 50% CI prevalence across all simulations that reach x detected cases.
}
\keyword{datasets}

