% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_zika_outbreaks.R
\name{run_n_zika_sims}
\alias{run_n_zika_sims}
\title{Run multiple zika simulations}
\usage{
run_n_zika_sims(num_reps, ...)
}
\arguments{
\item{num_reps}{An integer.}

\item{...}{A list of parameters to call run_zika_sim.}
}
\value{
A list of length num_reps, where each component is a single call to run_zika_sim.
}
\description{
A function to run multiple zika simulations with the same parameters
}
\examples{
run_n_zika_sims(1, zika_def_parms())
}

