% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_zika_outbreaks.R
\name{run_zika_sim}
\alias{run_zika_sim}
\title{Runs a single zika simulation}
\usage{
run_zika_sim(params)
}
\arguments{
\item{params}{List of named parameters for simulation, starred parameters must be present, but suggested to call zika_def_parms() and alter parms according to specific run (See examples).}

\item{infBoxes*}{Integer specifying number of infectious class boxes}

\item{incBoxes*}{Integer specifying number of incubation class boxes}

\item{recov_p*}{Daily probability of recovery for each infectious box}

\item{incub_rate*}{Daily probability of transitioning for each incubation box}

\item{prop_p*}{Daily Poisson rate of transmission for each infectious individual}

\item{e_thresh*}{Maximum number of cumulative infections before sim ends}

\item{prob_symp*}{Probability that cases are symptomatic (haven't tested anything <1)}

\item{dis_prob_symp*}{Probability of discovery for symptomatic individuals}

\item{dis_prob_asymp*}{Probability of discovery for asymptomatic individuals}

\item{intro_rate*}{Daily rate of new introductions according to poisson distribution}
}
\value{
Returns a dataframe that contains the daily evolution of the model compartments
}
\description{
Runs a single stochastic branching process zika simulation according to the model description
in "Real-time Zika Risk assessment in the United States."
}
\examples{
run_zika_sim(zika_def_parms())
run_zika_sim(zika_def_parms(r_not=0.8, intro_rate=0.1))
}

