% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{texas_county_shp}
\alias{texas_county_shp}
\title{Texas county map}
\format{A spatial dataset for all 254 Texas counties
\describe{
\item{counties}{Texas shapefile object, used to plot maps}
}}
\source{
online
}
\usage{
texas_county_shp
}
\description{
Spatial data for all counties in Texas, can be fortified to turn into dataframe
}
\keyword{datasets}
