% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_zika_outbreaks.R
\name{track_infectious}
\alias{track_infectious}
\title{Track movement through infectious classes}
\usage{
track_infectious(UI, recov_prob, disc_prob)
}
\arguments{
\item{UI}{Integer vector length=number of boxes}

\item{recov_prob}{Probability of recovery}

\item{disc_prob}{Probability of being reported}
}
\value{
Outputs a list UI recovering, UI reporting, and amount to add to each DI box
}
\description{
A function that tracks movement between undetected and detected classes of individuals.
Progresses individuals according to one step (day), and accounts for the fact that individuals
can recover and be reported in the same time period
}
\examples{
track_infectious(c(10,50,20), 0.3, 0.5)
}

