% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_zika_outbreaks.R
\name{update_state}
\alias{update_state}
\title{Update state movement}
\usage{
update_state(n, prob)
}
\arguments{
\item{n}{An integer number of draws.}

\item{prob}{A probability of transitioning.}
}
\value{
Number of individuals leaving state.
}
\description{
A function that calculates how many individuals transition out of a state
}
\examples{
update_state(100, 0.3)
}

