% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_zika_outbreaks.R
\name{update_states}
\alias{update_states}
\title{Update multiple state's movement}
\usage{
update_states(n, prob)
}
\arguments{
\item{n}{An integer vector for the number of draws for each state.}

\item{prob}{A probability of transitioning.}
}
\value{
Number of individuals leaving each state.
}
\description{
A vectorized function of update_state()
}
\examples{
update_states(c(50,100,20,30), 0.3)
}

