% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_zika_outbreaks.R
\name{zika_def_parms}
\alias{zika_def_parms}
\title{Default Zika parameters for sims}
\usage{
zika_def_parms(r_not = 1.1, infBoxes = 3, incBoxes = 6,
  recov_p = 0.3040571/(3/infBoxes), incub_rate = 0.583917, prop_p = r_not
  * recov_p/infBoxes, e_thresh = 100, prob_symp = 1, dis_prob_symp = 0.01,
  dis_prob_asymp = 0, intro_rate = 0.01)
}
\arguments{
\item{r_not}{value specifying outbreak R0, which automatically modulates transmission rate}

\item{infBoxes}{Integer specifying number of infectious class boxes}

\item{incBoxes}{Integer specifying number of incubation class boxes}

\item{recov_p}{Daily probability of recovery for each infectious box}

\item{incub_rate}{Daily probability of transitioning for each incubation box}

\item{prop_p}{Daily Poisson rate of transmission for each infectious individual}

\item{e_thresh}{Maximum number of cumulative infections before sim ends}

\item{prob_symp}{Probability that cases are symptomatic (haven't tested anything <1)}

\item{dis_prob_symp}{Probability of discovery for symptomatic individuals}

\item{dis_prob_asymp}{Probability of discovery for asymptomatic individuals}

\item{intro_rate}{Daily rate of new introductions according to poisson distribution}
}
\value{
A list of length num_reps, where each component is a single call to run_zika_sim.
}
\description{
A function that gives a list of parameters for running a zika simulation.
Defaults to parameters for Zika with no introductions.
}
\examples{
zika_def_parms()
zika_def_parms(r_not=1.6, intro_rate=0.1)
}

