% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_triggers_data}
\alias{exp_triggers_data}
\title{County surveillance triggers}
\format{An object of class \code{data.frame} with 149 rows and 9 columns.}
\source{
county simulation data 50% probability of an epidemic
}
\usage{
exp_triggers_data
}
\description{
Figure 4c data - number of reported cases necessary for each county to have 50% probability of an epidemic
}
\keyword{datasets}

