% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_zika_outbreaks.R
\name{get_county_parms}
\alias{get_county_parms}
\title{Get County-specific parameters}
\usage{
get_county_parms(county, rnot_bound = "median", ...)
}
\arguments{
\item{county}{Integer or character string identifying either the county id number or the county name#' @param ... Function can take any other arguments to zika_def_parms}

\item{rnot_bound}{string specifying which Rnot is desired ("lower", "median", or "upper")}
}
\value{
A list of length num_reps, where each component is a single call to run_zika_sim.
}
\description{
A function that finds the correct parameters for a specified county to be used for zika outbreak simulations
}
\examples{
get_county_parms("travis")
get_county_parms(15)
get_county_parms("travis", e_thresh=1000)
}
